requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
itsdangerous==2.0.1
click==8.0.1
Jinja2==3.0.1
python-dotenv==0.19.2
pandas==1.3.5
scikit-learn==1.3.2
flask-wtf==1.0.0
plotly==5.5.0
gunicorn==20.1.0
nvdlib==0.5.0
